+++ roles/ansible-hugo/.gitignore
@@ -1,0 +1,9 @@
+.vagrant
+*.retry
+VAULT_PASSWORD
+VAULT_PASS
+.vault_pass
+.vault_pass.asc
+vagramt/fetch
+vagrant/ubuntu-*.log
+__pycache__
