+++ roles/ansible-hugo/.travis.yml
@@ -1,0 +1,54 @@
+---
+
+# Thanks to geerlingguy!
+
+language: python
+services: docker
+
+env:
+  global:
+    - ROLE_NAME: hugo
+  matrix:
+    # Local user install
+    - MOLECULE_DISTRO: centos7
+    - MOLECULE_DISTRO: centos8
+    - MOLECULE_DISTRO: debian10
+    - MOLECULE_DISTRO: ubuntu1804
+    - MOLECULE_DISTRO: fedora29
+    - MOLECULE_DISTRO: amazonlinux2
+
+    # Global install
+    - MOLECULE_DISTRO: centos7
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+    - MOLECULE_DISTRO: centos8
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+    - MOLECULE_DISTRO: debian10
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+    - MOLECULE_DISTRO: ubuntu1804
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+    - MOLECULE_DISTRO: fedora29
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+    - MOLECULE_DISTRO: amazonlinux2
+      MOLECULE_PREPARE_PLAYBOOOK: prepare-global.yml
+      MOLECULE_PLAYBOOK: playbook-global.yml
+install:
+  # Install test dependencies.
+  - pip install molecule==2.22 docker
+
+before_script:
+  # Use actual Ansible Galaxy role name for the project directory.
+  - cd ../
+  - mv ansible-role-$ROLE_NAME xanmanning.$ROLE_NAME
+  - cd xanmanning.$ROLE_NAME
+
+script:
+  # Run tests.
+  - molecule test
+
+notifications:
+  webhooks: https://galaxy.ansible.com/api/v1/notifications/
