+++ roles/ansible-hugo/defaults/main.yml
@@ -1,12 +1,16 @@
 ---
-# defaults file for hugo
 
-hugo_app: hugo_extended
-hugo_version: 0.92.2
-hugo_os: "{{ ansible_system }}"
-hugo_arch: "64bit"
-hugo_dl_url: https://github.com/gohugoio/hugo/releases/download/v{{ hugo_version }}/{{ hugo_app }}_{{ hugo_version }}_{{ hugo_os }}-{{ hugo_arch }}.tar.gz
-hugo_bin_path: /usr/local/bin
-hugo_file_owner: root
-hugo_file_group: root
-hugo_file_mode: '0755'
+# Use a specific version of Hugo, eg. `0.66.0`. Specify `false` for latest.
+hugo_version: false
+
+# Allow role to install OS dependencies.
+hugo_install_os_dependencies: false
+
+# Installation directory for Hugo.
+hugo_install_dir: "{{ ansible_user_dir }}/bin"
+
+# Directory to put Hugo projects. Specify `false` to skip.
+hugo_projects_dir: "{{ ansible_user_dir }}/projects"
+
+# List of Hugo projects to be cloned with `git`. See notes in README.md
+# hugo_projects: []
