+++ roles/ansible-hugo/handlers/main.yml
@@ -1,0 +1,2 @@
+---
+# handlers file for ansible-role-hugo
