+++ roles/ansible-hugo/meta/main.yml
@@ -1,30 +1,82 @@
 ---
 galaxy_info:
   role_name: hugo
-  author: darkwizard242
-  description: Installs/Uninstalls 'Hugo', which is one of the most popular open source static site generator.
-  company: none
-  license: MIT
-  min_ansible_version: 2.4
+  author: Xan Manning
+  description: Ansible Role to install Hugo for building websites
+  company: Pyrat Ltd.
 
+  # If the issue tracker for your role is not on github, uncomment the
+  # next line and provide a value
+  # issue_tracker_url: http://example.com/issue/tracker
+
+  # Some suggested licenses:
+  # - BSD (default)
+  # - MIT
+  # - GPLv2
+  # - GPLv3
+  # - Apache
+  # - CC-BY
+  license: BSD
+
+  min_ansible_version: 2.7
+
+  # If this a Container Enabled role, provide the minimum Ansible Container version.
+  # min_ansible_container_version:
+
+  # Optionally specify the branch Galaxy will use when accessing the GitHub
+  # repo for this role. During role install, if no tags are available,
+  # Galaxy will use this branch. During import Galaxy will access files on
+  # this branch. If Travis integration is configured, only notifications for this
+  # branch will be accepted. Otherwise, in all cases, the repo's default branch
+  # (usually master) will be used.
+  # github_branch:
+
+  #
+  # platforms is a list of platforms, and each platform has a name and a list of versions.
+  #
   platforms:
-    - name: Ubuntu
-      versions:
-        - bionic
-        - focal
-    - name: Debian
-      versions:
-        - stretch
-        - buster
     - name: EL
       versions:
         - 7
         - 8
+    - name: Amazon
+    - name: Fedora
+      versions:
+        - 29
+        - 30
+        - 31
+    - name: Debian
+      versions:
+        - buster
+    - name: SLES
+      versions:
+        - 15
+    - name: Ubuntu
+      versions:
+        - bionic
+  # - name: SomePlatform
+  #   versions:
+  #   - all
+  #   - 1.0
+  #   - 7
+  #   - 99.99
 
   galaxy_tags:
+    - development
+    - golang
     - hugo
-    - system
-    - os
-    - staticsitegenerator
-    - framework
-    - frontend
+    - websites
+    - static
+    - html
+    - css
+    - cloud
+# List tags for your role here, one per line. A tag is a keyword that describes
+# and categorizes the role. Users find roles by searching for tags. Be sure to
+# remove the '[]' above, if you add tags to this list.
+#
+# NOTE: A tag is limited to a single word comprised of alphanumeric characters.
+#       Maximum 20 tags per role.
+
+dependencies: []
+# List your role dependencies here, one per line. Be sure to remove the '[]' above,
+# if you add dependencies to this list.
