+++ roles/ansible-hugo/molecule/default/playbook.yml
@@ -1,0 +1,7 @@
+---
+- name: Converge
+  hosts: all
+  become: true
+  become_user: hugouser
+  roles:
+    - role: xanmanning.hugo
