+++ roles/ansible-hugo/molecule/default/prepare-global.yml
@@ -1,0 +1,15 @@
+---
+
+- name: Prepare containers for global install testing
+  hosts: all
+  become: true
+  tasks:
+    - name: Ensure apt cache is updated
+      apt:
+        update_cache: true
+      when: ansible_pkg_mgr == 'apt'
+
+    - name: Ensure tar is installed
+      package:
+        name: tar
+        state: present
