+++ roles/ansible-hugo/molecule/default/prepare.yml
@@ -1,0 +1,26 @@
+---
+
+- name: Prepare containers for local-user install testing
+  hosts: all
+  become: true
+  tasks:
+    - name: Ensure a user group exists
+      group:
+        name: user
+        state: present
+
+    - name: Ensure a normal user exists
+      user:
+        name: hugouser
+        group: user
+        state: present
+
+    - name: Ensure apt cache is updated
+      apt:
+        update_cache: true
+      when: ansible_pkg_mgr == 'apt'
+
+    - name: Ensure tar is installed
+      package:
+        name: tar
+        state: present
