+++ roles/ansible-hugo/molecule/default/tests/test_default.py
@@ -6,25 +6,9 @@ testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
     os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
 
 
-PACKAGE_BINARY = '/usr/local/bin/hugo'
+def test_hosts_file(host):
+    f = host.file('/etc/hosts')
 
-
-def test_hugo_binary_exists(host):
-    """
-    Tests if hugo binary exists.
-    """
-    assert host.file(PACKAGE_BINARY).exists
-
-
-def test_hugo_binary_file(host):
-    """
-    Tests if hugo binary is a file type.
-    """
-    assert host.file(PACKAGE_BINARY).is_file
-
-
-def test_hugo_binary_which(host):
-    """
-    Tests the output to confirm hugo's binary location.
-    """
-    assert host.check_output('which hugo') == PACKAGE_BINARY
+    assert f.exists
+    assert f.user == 'root'
+    assert f.group == 'root'
