+++ roles/ansible-hugo/tasks/debian/install-os-dependencies.yml
@@ -1,0 +1,6 @@
+---
+
+- name: Ensure git is installed
+  apt:
+    name: git
+    state: present
