+++ roles/ansible-hugo/tasks/main.yml
@@ -1,10 +1,20 @@
 ---
-# tasks file for hugo
 
-- name: Import install_debian.yml if OS family is Debian
-  import_tasks: install_debian.yml
-  when: ansible_os_family == "Debian"
+- import_tasks: pre-checks.yml
 
-- name: Import install_el.yml if OS family is EL
-  import_tasks: install_el.yml
-  when: ansible_os_family == "RedHat"
+- include_tasks: "{{ ansible_os_family | lower }}/install-os-dependencies.yml"
+  when: hugo_install_os_dependencies is defined
+        and hugo_install_os_dependencies
+
+- import_tasks: get-version.yml
+  when: hugo_version is not defined or not hugo_version
+
+- import_tasks: download-hugo.yml
+
+- import_tasks: install-hugo.yml
+
+- import_tasks: download-projects.yml
+  when: go_projects is defined
+        and go_projects | length >= 1
+
+- import_tasks: post-checks.yml
