+++ roles/ansible-hugo/tasks/redhat/install-os-dependencies.yml
@@ -1,0 +1,13 @@
+---
+
+- name: Ensure git is installed (YUM)
+  yum:
+    name: git
+    state: present
+  when: ansible_pkg_mgr == "yum"
+
+- name: Ensure git is installed (DNF)
+  dnf:
+    name: git
+    state: present
+  when: ansible_pkg_mgr == "dnf"
