+++ roles/ansible-hugo/tasks/suse/install-os-dependencies.yml
@@ -1,0 +1,6 @@
+---
+
+- name: Ensure git is installed
+  zypper:
+    name: git
+    state: present
