--- roles/hugo_custom.yml
@@ -1,17 +1,0 @@
----
-- hosts: sorrowset-vm-1
-  become: true
-  vars:
-    hugo_install_dir: /usr/local/bin
-    hugo_projects:
-        - name: sludge.one
-          repo: 
-          key_file: "{{ ansible_user_dir }}/devops.pub"
-  roles: 
-    - role: ansible-hugo
-
-      
-
-
-# Same role diff user name it seems, or forked
-# - role: xanmanning.hugo
