+++ roles/not.ansible-hugo/.yamllint
@@ -1,4 +1,8 @@
+---
 extends: default
+
+ignore: |
+  .github/
 
 rules:
   braces:
