+++ roles/not.ansible-hugo/defaults/main.yml
@@ -1,0 +1,12 @@
+---
+# defaults file for hugo
+
+hugo_app: hugo_extended
+hugo_version: 0.92.2
+hugo_os: "{{ ansible_system }}"
+hugo_arch: "64bit"
+hugo_dl_url: https://github.com/gohugoio/hugo/releases/download/v{{ hugo_version }}/{{ hugo_app }}_{{ hugo_version }}_{{ hugo_os }}-{{ hugo_arch }}.tar.gz
+hugo_bin_path: /usr/local/bin
+hugo_file_owner: root
+hugo_file_group: root
+hugo_file_mode: '0755'
