+++ roles/not.ansible-hugo/molecule/default/tests/test_default.py
@@ -1,0 +1,30 @@
+import os
+
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
+
+PACKAGE_BINARY = '/usr/local/bin/hugo'
+
+
+def test_hugo_binary_exists(host):
+    """
+    Tests if hugo binary exists.
+    """
+    assert host.file(PACKAGE_BINARY).exists
+
+
+def test_hugo_binary_file(host):
+    """
+    Tests if hugo binary is a file type.
+    """
+    assert host.file(PACKAGE_BINARY).is_file
+
+
+def test_hugo_binary_which(host):
+    """
+    Tests the output to confirm hugo's binary location.
+    """
+    assert host.check_output('which hugo') == PACKAGE_BINARY
