+++ roles/not.ansible-hugo/sonar-project.properties
