+++ roles/not.ansible-hugo/tasks/main.yml
@@ -1,0 +1,10 @@
+---
+# tasks file for hugo
+
+- name: Import install_debian.yml if OS family is Debian
+  import_tasks: install_debian.yml
+  when: ansible_os_family == "Debian"
+
+- name: Import install_el.yml if OS family is EL
+  import_tasks: install_el.yml
+  when: ansible_os_family == "RedHat"
